requirements.txt
alembic==1.8.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
greenlet==2.0.1
Mako==1.2.3
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
Pygments==2.13.0
python-dateutil==2.8.2
rich==12.6.0
shellingham==1.5.0
six==1.16.0
SQLAlchemy==2.0.0b3
termcolor==2.1.0
typer==0.7.0
typing_extensions==4.4.0
